LGPL 2.0 is long out of date, AGPL 3.0 is cool
IANAL, but GNU's L(ibrary) GPL v2.0 is way out of date. On the outlined upgrade path, all GPL license are fair to upgrade to from LGPL 2.0 from newer LGPL variants to the most stringle AGPL versions.  In the project's original spirit, I am eager to provide my updates as highly-in-line with copyleft.